requirements.txt
Django
python-magic==0.4.6
